Update typos
After fixing the the init files I just followed the chain of errors in fixing typos in various files to correctly reference tables and table properties.

for example user.PID and user.pid are not the same, and the property is all lowercase so previous entries of user.PID were fixed.

similarly table name refs to comments to comments, users to user, etc...