Backend post get wip (#25)
* Setup backend framework

* Ad get in post service and implement api /get for posts

I reference our ex04 to add the get function in post service and add the api get for posts to set up the backend part. This commit finished the main part of backend task 1

* update typos causing errors

changed post to Post

also combined the two seperate get methods that existed for some reason

post.get works but creatinig a post does not. i will fix this in post entity andd post service.

* Update post_entity.py

postEntity switches the PID of the user and does not return the actual user object.

also inn order to view the registered user we need to use the routes given in the /docs folder from kris jordan

---------

Co-authored-by: GuningShen <gunings@email.unc.edu>
Co-authored-by: Wenjing Huang <rubyhwj@email.unc.edu>