try to pass the unit test of post (#38)
* try to pass the unit test of post

* fix the error of post pytest1 of getting posts

The original problem is the AttributeError: 'Depend' object has no attribute 'scalar'. So after googling and asking chatgpt, I realized the problem wass the session is not actually type session but depend in post_servise. So Iinsteasd of using _session, which was actually depend, I created another session and put it in the post_servoce. And it worked! However, then another error came out saying 'post.title' column does not exist. I tried modify the entity and the model file but it did not fix anything. So I just comment out the title and the description attribute for now.

* Fix create post test bug

---------

Co-authored-by: Wenjing Huang <rubyhwj@email.unc.edu>