New front (#81)
* backend

* try to pass the unit test of post

* fix the error of post pytest1 of getting posts

The original problem is the AttributeError: 'Depend' object has no attribute 'scalar'. So after googling and asking chatgpt, I realized the problem wass the session is not actually type session but depend in post_servise. So Iinsteasd of using _session, which was actually depend, I created another session and put it in the post_servoce. And it worked! However, then another error came out saying 'post.title' column does not exist. I tried modify the entity and the model file but it did not fix anything. So I just comment out the title and the description attribute for now.

* Fix create post test bug

* Add sample posts for view

* new_look

* Merge branch 'stage' of https://github.com/comp423-23s/final-project-final-d4 into stage

* modified tag formate and add private comment to the frontend

* connected comments with the backend

* connected search with the backend

* connected create-post to backend

* complete create post

* added documentation files to docs

* frontend docstring

* new_1

* bug on connect comment

* try

* fixed comments bug

* adjust the get and post comment, still have bug in post comments

* still need to fix the create comment bug

---------

Co-authored-by: Wenjing Huang <rubyhwj@email.unc.edu>
Co-authored-by: GuningShen <gunings@email.unc.edu>